+++ scripts/start.js
@@ -1,3 +1,5 @@
+require('babel-core/register');
+
 process.env.NODE_ENV = 'development';
 
 const path = require('path');