+++ src/common/components/Tooltip/index.js
@@ -3,18 +3,17 @@
 import { Component } from 'react';
 import { connect } from 'react-redux';
 import { createSelector } from 'reselect';
-
-import styles from './styles.less';
+import cx from 'classnames';
 
 import { showTooltip } from 'features/Gw2/actions';
-
+import ArmoryBadge from 'common/components/ArmoryBadge';
 import MouseFollow from '../MouseFollow';
-
 import AmuletTooltip from './Amulet';
 import ItemTooltip from './Item';
 import SkillTooltip from './Skill';
 import SimpleTooltip from './Simple';
 import Background from './Background';
+import styles from './styles.less';
 
 const selector = createSelector(
   (state) => state.tooltip,
@@ -24,25 +23,30 @@ const selector = createSelector(
 );
 
 type Props = {
-  tooltip: {
+  tooltip?: {
     show: boolean,
     type: string,
     data: Object,
   },
-  dispatch: Function,
+  showTooltip?: () => void,
+  showBadge?: boolean,
+  className?: string,
 };
 
-class Tooltip extends Component {
+@connect(selector, {
+  showTooltip,
+})
+export default class Tooltip extends Component {
   props: Props;
 
   close = () => {
-    this.props.dispatch(showTooltip(false));
+    this.props.showTooltip && this.props.showTooltip(false);
   };
 
   render () {
-    const { tooltip } = this.props;
+    const { tooltip, showBadge, className } = this.props;
 
-    if (!tooltip.show) return null;
+    if (!tooltip || !tooltip.show) return null;
 
     let content;
 
@@ -67,12 +71,11 @@ class Tooltip extends Component {
 
     return (
       <MouseFollow onTouchEnd={this.close}>
-        <div className={styles.root}>
+        <div className={cx(styles.root, className)}>
           {content}
+          {showBadge && <ArmoryBadge />}
         </div>
       </MouseFollow>
     );
   }
 }
-
-export default connect(selector)(Tooltip);