+++ src/common/components/TooltipTrigger/index.js
@@ -5,27 +5,36 @@ import { connect } from 'react-redux';
 import { showTooltip } from 'features/Gw2/actions';
 
 type Props = {
-  data: string | Object,
-  dispatch: Function,
-  children: React$Element<*>,
-  type: string,
+  data?: string | Object,
+  showTooltip?: Function,
+  children?: React$Element<*>,
+  type?: string,
 };
 
-class TooltipTrigger extends Component {
+@connect(null, {
+  showTooltip,
+})
+export default class TooltipTrigger extends Component {
   props: Props;
 
   showTooltip = () => {
-    this.props.dispatch(showTooltip(true, {
+    const data = {
       data: this.props.data,
       type: this.props.type,
-    }));
+    };
+
+    this.props.showTooltip && this.props.showTooltip(true, data);
   };
 
   hideTooltip = () => {
-    this.props.dispatch(showTooltip(false));
+    this.props.showTooltip && this.props.showTooltip(false);
   };
 
   render () {
+    if (!this.props.children) {
+      return null;
+    }
+
     return cloneElement(this.props.children, {
       onMouseEnter: this.showTooltip,
       onMouseLeave: this.hideTooltip,
@@ -33,5 +42,3 @@ class TooltipTrigger extends Component {
     });
   }
 }
-
-export default connect()(TooltipTrigger);