+++ src/embeds/index.js
@@ -1,0 +1,14 @@
+// @flow
+
+// $FlowFixMe
+const load = require('promise?global!./bootstrap'); // eslint-disable-line import/no-webpack-loader-syntax
+
+function init () {
+  load().then(({ default: bootstrap }) => bootstrap());
+}
+
+if (document.readyState === 'loading') {
+  document.addEventListener('DOMContentLoaded', init);
+} else {
+  setTimeout(init, 1);
+}