+++ src/embeds/items.js
@@ -1,0 +1,7 @@
+// @flow
+
+import Items from './components/Items';
+
+export default function (element: HTMLElement, ids: Array<number>) {
+  return () => <Items ids={ids} />;
+}