+++ src/embeds/skills.js
@@ -1,0 +1,7 @@
+// @flow
+
+import Skills from './components/Skills';
+
+export default function (element: HTMLElement, ids: Array<number>) {
+  return () => <Skills ids={ids} />;
+}