+++ src/index.js
@@ -2,7 +2,7 @@
 
 import ReactDOM from 'react-dom';
 import Router from './Router';
-import Base from './base';
+import Base from './Base';
 
 ReactDOM.render(
   <Base>